requirements.txt
uvicorn
fastapi
google-cloud-speech
vertexai
google-cloud-aiplatform
IPython
openai
elevenlabs
streamlit